feat: enable CSV injector in NamespaceScope CR (#92)
* feat: enable CSV injector in NamespaceScope CR

In order to make the NamespaceScope operator can be used in more situations, it should have the capability to patch CSVs of Operators to convert them can be managed by the NamespaceScope operator.

"csvInjector.enable" is used to control if the NamespaceScope operator will patch the CSVs of specific operators. The default value is false.

* Fix typo